requirements.txt
Flask==3.0.2
gunicorn==21.2.0
google-genai==0.3.0
